requirements.txt
google-generativeai
pandas
PyMuPDF